+++ lib/src/testainers_base.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:testainers/src/testainers_config.dart';
@@ -16,6 +17,12 @@ abstract class Testainers {
   final Map<String, String> env;
   final bool detached;
   final bool remove;
+  final String healthCmd;
+  final int healthInterval;
+  final int healthRetries;
+  final int healthTimeout;
+  final int healthStartPeriod;
+  final bool noHealthcheck;
   String? id;
 
   ///
@@ -28,6 +35,12 @@ abstract class Testainers {
     required this.env,
     required this.detached,
     required this.remove,
+    this.healthCmd = '',
+    this.healthInterval = 0,
+    this.healthRetries = 0,
+    this.healthTimeout = 0,
+    this.healthStartPeriod = 0,
+    this.noHealthcheck = false,
     this.config = const TestainersConfig(),
     String? name,
   }) : name = name ?? TestainersUtils.generateName();
@@ -78,15 +91,43 @@ abstract class Testainers {
     for (final MapEntry<String, String> entry in effectiveEnv.entries) {
       arguments
         ..add('-e')
-        ..add('"${entry.key}=${entry.value}"');
+        ..add('${entry.key}=${entry.value}');
+    }
+
+    if (healthCmd.isNotEmpty) {
+      arguments.add('--health-cmd=$healthCmd');
+
+      if (healthInterval > 0) {
+        arguments.add('--health-interval=${healthInterval}s');
+      }
+
+      if (healthRetries > 0) {
+        arguments.add('--health-retries=$healthRetries');
+      }
+
+      if (healthTimeout > 0) {
+        arguments.add('--health-timeout=${healthTimeout}s');
+      }
+
+      if (healthStartPeriod > 0) {
+        arguments.add('--health-start-period=${healthStartPeriod}s');
+      }
+    }
+
+    if (noHealthcheck) {
+      arguments.add('--no-healthcheck');
     }
 
     arguments.add('$image:$tag');
 
-    // print('${config.runner} ${arguments.join(' ')}');
+    if (config.debug) {
+      print('${config.runner} ${arguments.join(' ')}');
+    }
 
-    result =
-        await Process.run(config.runner, arguments).timeout(config.timeout);
+    result = await Process.run(
+      config.runner,
+      arguments,
+    ).timeout(config.timeout);
 
     if (result.exitCode != 0) {
       throw TestainersException(
@@ -97,9 +138,51 @@ abstract class Testainers {
 
     id = result.stdout.toString().trim();
 
-    // TODO(edufolly): Use healthcheck to continue.
-    // https://docs.docker.com/engine/reference/run/#healthcheck
-    await Future<void>.delayed(bootSleep ?? config.bootSleep);
+    if (healthCmd.isEmpty || noHealthcheck) {
+      await Future<void>.delayed(bootSleep ?? config.bootSleep);
+    } else {
+      late Map<dynamic, dynamic> inspectMap;
+
+      String healthStatus = 'starting';
+
+      while (healthStatus == 'starting') {
+        await Future<void>.delayed(const Duration(seconds: 1));
+
+        result = await Process.run(
+          config.runner,
+          <String>[
+            'inspect',
+            "--format='{{json .State.Health}}'",
+            name,
+          ],
+        ).timeout(config.timeout);
+
+        if (result.exitCode != 0) {
+          throw TestainersException(
+            'Container $name with image $image:$tag could not be inspected.',
+            reason: result.stderr,
+          );
+        }
+
+        final String inspectString = result.stdout.toString().trim();
+
+        inspectMap =
+            jsonDecode(inspectString.substring(1, inspectString.length - 1));
+
+        if (config.debug) {
+          print('Inspect: $inspectMap');
+        }
+
+        healthStatus = inspectMap['Status'];
+      }
+
+      if (healthStatus != 'healthy') {
+        throw TestainersException(
+          'Container $name with image $image:$tag is $healthStatus.',
+          reason: inspectMap,
+        );
+      }
+    }
   }
 
   ///