+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2 - 2023-06-30
+
+- Adding Mongodb container.
+- Adding healthcheck feature.
+- Creating CI scripts.
+
 ## 0.0.1 - 2023-06-29
 
 - Initial version.